Introduce warning for legacy engine version
Due to marketplace restrictions we can not update the plugin for engine
versions older than the latest three. We can however leave an "old"
version of the SDK up there for visibility and this should direct you to
use GitHub instead. This change is a special release and should never be
merged to main